[FLINK-28678][table-planner] Copy some nodes from calcite, and need revert when upgrade calcite-1.31
Since currently the feat [CALCITE-5107] is not introduced to the calcite version flink used, so just temporarily copy the files and cherry-pick this feat directly. When the calcite upgrades to 1.31, these classes should be deleted.

This pr is just to copy some nodes from calcite, but doesn't do any modification on them.

This closes #20359