[FLINK-36279][runtime] Making desired resources being calculated based on all available slots for the job
FLINK-36014 aligned the triggering of the CreateExecutionGraph state within WaitingForResources and Executing state. Before that change, only WaitingForResources relied on this method. Relying on free slots was good enough because in WaitingForResources state, there are no slots allocated, yet.

Using this method for Executing state now as well changes this premise because there are slots allocated while checking the slot availability that would become available after the restart. Hence, considering these slots in the slot availability check is good enough. This will not break the premise for the WaitingForResources state.